+++ Demo/Sources/Demos/DemosRootViewController.swift
@@ -143,6 +143,14 @@ public final class DemosRootViewController : UIViewController
                     DemoItem(text: "Tappable Row"),
                     selectionStyle: .tappable
                 )
+                
+                section += Item(
+                    DemoItem(text: "Tappable Row (Slow Is Selected)"),
+                    selectionStyle: .tappable,
+                    onSelect: { _ in
+                        Thread.sleep(forTimeInterval: 0.5)
+                    }
+                )
             }
             
             list += Section(identifier: "collection-view") { section in