+++ Listable/Sources/Internal/Presentation State/PresentationState.ItemState.swift
@@ -277,10 +277,27 @@ extension PresentationState
         {
             self.storage.state.isSelected = isSelected
             
-            if isSelected {
-                self.model.onSelect?(self.model.content)
-            } else {
-                self.model.onDeselect?(self.model.content)
+            /// Schedule the caller-provided callbacks to happen after one runloop. Why?
+            ///
+            /// Because this method is called from within `UICollectionViewDelegate` callbacks,
+            /// This delay gives the `UICollectionView` time to schedule any necessary animations
+            /// for changes to the highlight and selection state – otherwise, these animations get
+            /// stuck behind the call to the `onSelect` or `onDeselect` blocks, which creates the appearance
+            /// of a laggy UI if these callbacks are slow.
+            DispatchQueue.main.async {
+                if isSelected {
+                    if let onSelect = self.model.onSelect {
+                        SignpostLogger.log(log: .listInteraction, name: "Item onSelect", for: self.model) {
+                            onSelect(self.model.content)
+                        }
+                    }
+                } else {
+                    if let onDeselect = self.model.onDeselect {
+                        SignpostLogger.log(log: .listInteraction, name: "Item onDeselect", for: self.model) {
+                            onDeselect(self.model.content)
+                        }
+                    }
+                }
             }
         }