+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Changed
 
+- Change `Item`'s `onSelect` and `onDeselect` [to be performed asynchronously](https://github.com/kyleve/Listable/pull/155) after a single runloop spin, to give `UICollectionView` time to schedule animations if these callbacks are slow. Add improved signpost logging for selection and deselection. 
+
 ### Misc
 
 # Past Releases